requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1

# AI SDKs
anthropic==0.25.0
openai==1.14.0
google-generativeai==0.4.0

# Notion API
notion-client==2.2.1

# Async HTTP
aiohttp==3.9.3

# Development Tools
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
black==24.2.0
ruff==0.3.0
